reorganized files into newly created organized folders. Also added individual styling files for each component inside of the Dashboard component/page
