+++ src/components/Dashboard/styles/Tabs-styling.js
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const StyledDiv = styled.div``;
