+++ src/components/Login-Signup/Signup.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import * as yup from 'yup';
-import { axiosWithAuth } from '../utils/axiosWithAuth';
+import { axiosWithAuth } from '../../utils/axiosWithAuth';
 import { StyledDiv } from './LoginSignup-styling';
 
 export default function Signup({ history }) {
